% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{call_api}
\alias{call_api}
\title{Make a call to the PriLIMS API}
\usage{
call_api(key, endpoint)
}
\arguments{
\item{key}{An API key used for accessing PriLIMS API}

\item{endpoint}{Endpoint URL; e.g. "raw" if accessing Raw data table.}
}
\description{
Requires an API key provided by Priogen Corp.
}
