% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{call_api}
\alias{call_api}
\title{Make a call to the PriLIMS API}
\usage{
call_api(
  key,
  endpoint,
  order = NULL,
  sample = NULL,
  reaction = NULL,
  local_test = FALSE
)
}
\arguments{
\item{key}{An API key used for accessing PriLIMS API}

\item{endpoint}{Endpoint URL; e.g. "results" if accessing Results data table.}

\item{order}{Character; the order number}

\item{sample}{Character; the sample name}

\item{reaction}{Character; the reaction name}

\item{local_test}{Logical; Are you testing on a local PriLIMS instance?}
}
\description{
Requires an API key provided by Priogen Corp.
}
